+++ display/index.js
@@ -3,6 +3,7 @@ const describeConfidence = require('./describe_confidence');
 const describeRoute = require('./describe_route');
 const describeRoutingFailure = require('./describe_routing_failure');
 const formatFeeRate = require('./format_fee_rate');
+const getIcons = require('./get_icons');
 const parseAmount = require('./parse_amount');
 const segmentMeasure = require('./segment_measure');
 const sumsForSegment = require('./sums_for_segment');
@@ -13,6 +14,7 @@ module.exports = {
   describeRoute,
   describeRoutingFailure,
   formatFeeRate,
+  getIcons,
   parseAmount,
   segmentMeasure,
   sumsForSegment,