+++ nodes/index.js
@@ -1,5 +1,11 @@
 const adjustSavedNodes = require('./adjust_saved_nodes');
+const adjustTags = require('./adjust_tags');
 const getSavedCredentials = require('./get_saved_credentials');
 const getSavedNodes = require('./get_saved_nodes');
 
-module.exports = {adjustSavedNodes, getSavedCredentials, getSavedNodes};
+module.exports = {
+  adjustSavedNodes,
+  adjustTags,
+  getSavedCredentials,
+  getSavedNodes,
+};